🏗 Move test store to main src package
